refactor(separate): Calculator 단위 테스트를 위해 추상화 및 책임 분리
