+++ src/main/java/calculator/Application.java
@@ -1,15 +1,13 @@
 package calculator;
 
-import calculator.domain.calculator.Calculator;
-import calculator.domain.calculator.Prompt;
-import java.util.List;
+import calculator.domain.Calculator;
 
 public class Application {
 
     public static void main(String[] args) {
-        Calculator calculator = new Calculator();
-        List<Integer> strings = calculator.extractInteger("1,2:3");
-        System.out.println(strings.toString());
+        Calculator calculator = new Calculator("1,2:3");
+        int sum = calculator.sum();
+        System.out.println(sum);
     }
 
 }
