--- src/main/java/calculator/domain/calculator/Calculator.java
@@ -1,57 +1,0 @@
-package calculator.domain.calculator;
-
-import java.util.List;
-
-public class Calculator {
-
-    private final Prompt prompt;
-
-    public Calculator() {
-        this.prompt = new Prompt();
-    }
-
-    public Integer sum(List<Integer> numbers) {
-        return numbers.stream()
-                .reduce(0, Integer::sum);
-    }
-
-    public List<Integer> extractInteger(String inputData) {
-        List<String> separatedData = prompt.separate(inputData);
-
-        return separatedData.stream()
-                .map(data -> {
-                    validateNumber(data);
-                    return Integer.valueOf(data);
-                })
-                .toList();
-    }
-
-    public List<Integer> extractInteger(String inputData, char customDelimiter) {
-        List<String> separatedData = prompt.separate(inputData, customDelimiter);
-
-        return separatedData.stream()
-                .map(data -> {
-                    validateNumber(data);
-                    return Integer.valueOf(data);
-                })
-                .toList();
-    }
-
-    private void validateNumber(String data) {
-        if (data == null || data.isEmpty()) {
-            throw new IllegalArgumentException();
-        }
-
-        for (char word: data.toCharArray()) {
-            if (this.isNotDigit(word)) {
-                throw new IllegalArgumentException();
-            }
-        }
-    }
-
-    private boolean isNotDigit(char word) {
-        return !Character.isDigit(word);
-    }
-
-
-}
